Change to Neptune logger. FEM multiprocessing improved
The local logging system is deprecated. Now logs are written to Neptune.
The FEM results are no longer processed in the sequence of complete. The received results are stored in the cache only, and transitions are resolved in the order of their step number.
The rewarding system is under testing and modification.